requirements.txt
requests
mangum
fastapi[standard]
boto3
bcrypt
pyjwt
google-auth
cachetools
aws-lambda-powertools